Edit cart and other improvements
Update so the program is now clearer, for example, tells the user what product has been selected and then asks them for a quantity. Other error where there was a input line where it should've been a print statement, better handling, removed a function that was no longer being used. Added edit cart function, and made a checkout function as well to reduce clutter in the car_menu() function.