Quantity and Delivery Refinement
User can no longer enter negative quantities, which would result in the store paying the user money, and there is confirmation on whether the user wants to proceed with the current delivery method, telling the user the price asking if they want to continue before asking for the address. Along with this, there is improved error handling for the postcode and phone number, where the program will check if the user has entered a phone number and post of the correct number of digits.